--- mrdc_ws/src/mrdc_example/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script_dir=$base/lib/mrdc_example
-[install]
-install_scripts=$base/lib/mrdc_example
