--- mrdc_ws/src/mrdc_example/src/subscriber.py
@@ -1,37 +1,0 @@
-# Copyright 2016 Open Source Robotics Foundation, Inc.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import rclpy
-
-from sensor_msgs.msg import Joy
-
-def main(args=None):
-    rclpy.init(args=args)
-
-    node = rclpy.create_node('minimal_subscriber')
-
-    subscription = node.create_subscription(Joy, '/joy', lambda msg: node.get_logger().info('I heard: "%s"' % msg), 10)
-    subscription  # prevent unused variable warning
-
-    rclpy.spin(node)
-
-    # Destroy the node explicitly
-    # (optional - otherwise it will be done automatically
-    # when the garbage collector destroys the node object)
-    node.destroy_node()
-    rclpy.shutdown()
-
-
-if __name__ == '__main__':
-    main()
