+++ mrdc_ws/src/mrdc_joy/launch/all_launch.xml
@@ -1,0 +1,4 @@
+<launch>
+    <node pkg="joy" exec="joy_node" name="joystick" />
+    <node pkg="mrdc_joy" exec="driver" name="drive" output="screen"/>
+</launch>
