+++ mrdc_ws/src/mrdc_joy/package.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>mrdc_example</name>
+  <name>mrdc_joy</name>
   <version>1.0.0</version>
-  <description>TODO: Package description</description>
+  <description>The MRDC Joy package that publishes controller input</description>
   <maintainer email="dylan.zemlin@gmail.com">Dylan Zemlin</maintainer>
   <license>Apache License 2.0</license>
