+++ mrdc_ws/src/mrdc_joy/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script_dir=$base/lib/mrdc_joy
+[install]
+install_scripts=$base/lib/mrdc_joy
