+++ mrdc_ws/src/mrdc_joy/setup.py
@@ -1,14 +1,14 @@
 from setuptools import find_packages, setup
 
-package_name = 'mrdc_example'
+package_name = 'mrdc_joy'
 
 setup(
     name=package_name,
     version='1.0.0',
     packages=find_packages(),
     data_files=[
-        ('share/ament_index/resource_index/packages', ['resource/' + package_name]),
         ('share/' + package_name, ['package.xml']),
+        ('share/' + package_name, ['launch/all_launch.xml'])
     ],
     install_requires=['setuptools'],
     maintainer='Dylan Zemlin',
@@ -17,8 +17,7 @@ setup(
     license='Apache License 2.0',
     entry_points={
         'console_scripts': [
-            'publisher = src.publisher:main',
-            'subscriber = src.subscriber:main'
+            'driver = src.driver:main'
         ],
     },
 )
